LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/coop-sandbox-redirects-cuts-opener.https-expected.txt

CONSOLE MESSAGE: Navigation was blocked by Cross-Origin-Opener-Policy
CONSOLE MESSAGE: Navigation was blocked by Cross-Origin-Opener-Policy

FAIL <iframe sandbox="allow-popups allow-scripts allow-same-origin"> Sandboxed Cross-Origin-Opener-Policy popup should cut the opener if necessary including during redirects. assert_equals: expected "true" but got "false"
FAIL <iframe sandbox="allow-popups allow-scripts"> Sandboxed Cross-Origin-Opener-Policy popup should cut the opener if necessary including during redirects. assert_equals: expected "true" but got "false"
